Remove `transformOutput` combinator
The fact that `transformOutput` required an `InvalidInputError` to be
explicitly created in order to model failure made this function
difficult to use (you had to keep track of your own `input`, etc). I
considered changing the signature to instead only require an error
message, but ultimately decided to ditch it entirely for now. Either
change is semver-major and I don't currently use `transformOutput` at
all (and I believe I'm still the sole consumer of this package).